pyproject.toml
[project]
name = "flet-chess"
version = "0.1.0"
description = "A two-player chess game built with Flet and custom SVG pieces."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flet[all]>=0.27.0",
    "python-chess>=1.999",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/.*",
    "main.py",  # UI code, hard to test
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "\\.\\.\\.",  # Protocol placeholder
]

[tool.coverage.branch]
fail_under = 100
